Exercise
The question is:
Create an array of Strings, each containing one of the top 10 reasons that you like Java. Prompt a user to enter a number from 1 to 10, convert the number to an integer, and then use the integer to display one of the reasons. 
( I used 5 reasons for checking, here you have to use parallell array if you go by the book and you have to use JOptionpane, used both arrays as a string but when I convert the array to int some problems were coming and also the elogic building a bit )