Partially complete exercise
The Question is:
A personal phone directory contains room for first names and phone 30 people. Assign names and phone numbers for the first 10 people. Prompt the user for a name and if the names is found in the list, display the corresponding phone. if the name is not found in the list, promote the user for a phone number; and the new name and phone number to the list. Continue to prompt the user for names until the user enter "quit". After the arrays are full (Containing 30 names), don not allow the user to add new entires.